moves the page creation stuff to utils
