+++ client/pages/programs.js
@@ -7,7 +7,7 @@ import {
   getAuthorizedProgramPolicies,
   decodeToken
 } from "global/utils/egoJwt";
-import { createPage } from "./_app";
+import { createPage } from "global/utils/pages";
 
 export default createPage({
   isPublic: true,
